+++ __tests__/sidebar/sidebar-modules.test.tsx
@@ -11,6 +11,7 @@ import {
   useSidebar,
 } from '@organisms/sidebar';
 
+
 function Wrapper() {
   return (
     <SidebarProvider>